Fix docstrings after splitting vim_status
